# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST in Golang with Gin Framework",
        "title": "User Routing Service",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/user": {
            "post": {
                "description": "Create available user from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Create Available User",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "user params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/objects.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responder.ResponseBody"
                        }
                    },
                    "400": {}
                }
            }
        },
        "/user/{id}": {
            "get": {
                "description": "Get available user from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Get Available User",
                "operationId": "get-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responder.ResponseBody"
                        }
                    }
                }
            },
            "put": {
                "description": "Update available user from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Update Available User",
                "operationId": "update-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user params",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/objects.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responder.ResponseBody"
                        }
                    },
                    "400": {}
                }
            },
            "delete": {
                "description": "Delete available user from db",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Delete Available User",
                "operationId": "delete-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responder.ResponseBody"
                        }
                    },
                    "400": {}
                }
            }
        }
    },
    "definitions": {
        "objects.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string",
                    "example": "服务端会自动生成"
                },
                "description": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "服务端会自动生成"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "responder.ResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}